makes submenu its own folder
